Trait cfg_traits::data::DataRegistry

pub trait DataRegistry<DataId, CollectionId> {
    type Collection: DataCollection<DataId, Data = Self::Data>;
    type Data;

    // Required methods
    fn get(
        data_id: &DataId,
        collection_id: &CollectionId
    ) -> Result<Self::Data, DispatchError>;
    fn collection(collection_id: &CollectionId) -> Self::Collection;
    fn register_id(
        data_id: &DataId,
        collection_id: &CollectionId
    ) -> DispatchResult;
    fn unregister_id(
        data_id: &DataId,
        collection_id: &CollectionId
    ) -> DispatchResult;
}

Abstraction that represents a storage where you can subscribe to data updates and collect them

Required Associated Types

type Collection: DataCollection<DataId, Data = Self::Data>

A collection of data

type Data

Represents a data

Required Methods

fn get( data_id: &DataId, collection_id: &CollectionId ) -> Result<Self::Data, DispatchError>

Return the last data value for a data id in a collection

fn collection(collection_id: &CollectionId) -> Self::Collection

Retrives a collection of data with all data associated to a collection id

fn register_id(data_id: &DataId, collection_id: &CollectionId) -> DispatchResult

Start listening data changes for a data id in a collection id

fn unregister_id( data_id: &DataId, collection_id: &CollectionId ) -> DispatchResult

Start listening data changes for a data id in a collection id