Trait cfg_traits::interest::InterestAccrual

pub trait InterestAccrual<Rate, Balance, Adjustment> {
    type MaxRateCount: Get<u32>;
    type NormalizedDebt: Member + Parameter + MaxEncodedLen + TypeInfo + Copy + Zero;
    type Rates: RateCollection<Rate, Balance, Self::NormalizedDebt>;

    // Required methods
    fn calculate_debt(
        interest_rate: &InterestRate<Rate>,
        normalized_debt: Self::NormalizedDebt,
        when: Seconds
    ) -> Result<Balance, DispatchError>;
    fn adjust_normalized_debt(
        interest_rate: &InterestRate<Rate>,
        normalized_debt: Self::NormalizedDebt,
        adjustment: Adjustment
    ) -> Result<Self::NormalizedDebt, DispatchError>;
    fn renormalize_debt(
        old_interest_rate: &InterestRate<Rate>,
        new_interest_rate: &InterestRate<Rate>,
        normalized_debt: Self::NormalizedDebt
    ) -> Result<Self::NormalizedDebt, DispatchError>;
    fn reference_rate(interest_rate: &InterestRate<Rate>) -> DispatchResult;
    fn unreference_rate(interest_rate: &InterestRate<Rate>) -> DispatchResult;
    fn validate_rate(interest_rate: &InterestRate<Rate>) -> DispatchResult;
    fn rates() -> Self::Rates;
}

A trait that can be used to calculate interest accrual for debt

Required Associated Types

type MaxRateCount: Get<u32>

The maximum number of rates this InterestAccrual can contain. It is necessary for rate calculations in consumers of this pallet, but is otherwise unused in this interface.

type NormalizedDebt: Member + Parameter + MaxEncodedLen + TypeInfo + Copy + Zero

type Rates: RateCollection<Rate, Balance, Self::NormalizedDebt>

Required Methods

fn calculate_debt( interest_rate: &InterestRate<Rate>, normalized_debt: Self::NormalizedDebt, when: Seconds ) -> Result<Balance, DispatchError>

Calculate the debt at an specific moment

fn adjust_normalized_debt( interest_rate: &InterestRate<Rate>, normalized_debt: Self::NormalizedDebt, adjustment: Adjustment ) -> Result<Self::NormalizedDebt, DispatchError>

Increase or decrease the normalized debt

fn renormalize_debt( old_interest_rate: &InterestRate<Rate>, new_interest_rate: &InterestRate<Rate>, normalized_debt: Self::NormalizedDebt ) -> Result<Self::NormalizedDebt, DispatchError>

Re-normalize a debt for a new interest rate

fn reference_rate(interest_rate: &InterestRate<Rate>) -> DispatchResult

Validate and indicate that a yearly rate is in use

fn unreference_rate(interest_rate: &InterestRate<Rate>) -> DispatchResult

Indicate that a rate is no longer in use

fn validate_rate(interest_rate: &InterestRate<Rate>) -> DispatchResult

Ask if the rate is valid to use by the implementation

fn rates() -> Self::Rates

Returns a collection of pre-computed rates to perform multiple operations with