Trait cfg_traits::investments::ForeignInvestment

pub trait ForeignInvestment<AccountId> {
    type Amount;
    type CurrencyId;
    type Error: Debug;
    type InvestmentId;

    // Required methods
    fn increase_foreign_investment(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        amount: Self::Amount,
        foreign_payment_currency: Self::CurrencyId,
        pool_currency: Self::CurrencyId
    ) -> Result<(), Self::Error>;
    fn decrease_foreign_investment(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        amount: Self::Amount,
        foreign_payment_currency: Self::CurrencyId,
        pool_currency: Self::CurrencyId
    ) -> Result<(), Self::Error>;
    fn increase_foreign_redemption(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        amount: Self::Amount,
        foreign_payout_currency: Self::CurrencyId
    ) -> Result<(), Self::Error>;
    fn decrease_foreign_redemption(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        amount: Self::Amount,
        foreign_payout_currency: Self::CurrencyId
    ) -> Result<(Self::Amount, Self::Amount), Self::Error>;
    fn collect_foreign_investment(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        foreign_currency: Self::CurrencyId
    ) -> Result<(), Self::Error>;
    fn collect_foreign_redemption(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        foreign_payout_currency: Self::CurrencyId,
        pool_currency: Self::CurrencyId
    ) -> Result<(), Self::Error>;
    fn investment(
        who: &AccountId,
        investment_id: Self::InvestmentId
    ) -> Result<Self::Amount, Self::Error>;
    fn redemption(
        who: &AccountId,
        investment_id: Self::InvestmentId
    ) -> Result<Self::Amount, Self::Error>;
    fn accepted_payment_currency(
        investment_id: Self::InvestmentId,
        currency: Self::CurrencyId
    ) -> bool;
    fn accepted_payout_currency(
        investment_id: Self::InvestmentId,
        currency: Self::CurrencyId
    ) -> bool;
}

Trait to handle investments in (presumably) foreign currencies, i.e., other currencies than the pool currency.

NOTE: Has many similarities with the Investment trait.

Required Associated Types

type Amount

type CurrencyId

type Error: Debug

type InvestmentId

Required Methods

fn increase_foreign_investment( who: &AccountId, investment_id: Self::InvestmentId, amount: Self::Amount, foreign_payment_currency: Self::CurrencyId, pool_currency: Self::CurrencyId ) -> Result<(), Self::Error>

Initiates the increment of a foreign investment amount in foreign_payment_currency of who into the investment class pool_currency to amount.

NOTE: In general, we can assume that the foreign and pool currencies mismatch and that swapping one into the other happens asynchronously. In that case, the finalization of updating the investment needs to be handled decoupled from the ForeignInvestment trait, e.g., by some hook.

fn decrease_foreign_investment( who: &AccountId, investment_id: Self::InvestmentId, amount: Self::Amount, foreign_payment_currency: Self::CurrencyId, pool_currency: Self::CurrencyId ) -> Result<(), Self::Error>

Initiates the decrement of a foreign investment amount in foreign_payment_currency of who into the investment class pool_currency to amount.

NOTE: In general, we can assume that the foreign and pool currencies mismatch and that swapping one into the other happens asynchronously. In that case, the finalization of updating the investment needs to be handled decoupled from the ForeignInvestment trait, e.g., by some hook.

fn increase_foreign_redemption( who: &AccountId, investment_id: Self::InvestmentId, amount: Self::Amount, foreign_payout_currency: Self::CurrencyId ) -> Result<(), Self::Error>

Initiates the increment of a foreign redemption amount for the given investment id.

NOTE: The foreign_payout_currency is only required to ensure subsequent redemption updating calls match to the original chosen foreign_payment_currency.

fn decrease_foreign_redemption( who: &AccountId, investment_id: Self::InvestmentId, amount: Self::Amount, foreign_payout_currency: Self::CurrencyId ) -> Result<(Self::Amount, Self::Amount), Self::Error>

Initiates the decrement of a foreign redemption amount.

NOTES:

The decrementing redemption amount is bound by the previously incremented redemption amount.

The foreign_payout_currency is only required for the potential dispatch of a response message.

fn collect_foreign_investment( who: &AccountId, investment_id: Self::InvestmentId, foreign_currency: Self::CurrencyId ) -> Result<(), Self::Error>

Collect the results of a user’s foreign invest orders for the given investment. If any amounts are not fulfilled they are directly appended to the next active order for this investment.

fn collect_foreign_redemption( who: &AccountId, investment_id: Self::InvestmentId, foreign_payout_currency: Self::CurrencyId, pool_currency: Self::CurrencyId ) -> Result<(), Self::Error>

Collect the results of a user’s foreign redeem orders for the given investment. If any amounts are not fulfilled they are directly appended to the next active order for this investment.

NOTE: The currency of the collected amount will be pool_currency whereas the user eventually wants to receive it in foreign_payout_currency.

fn investment( who: &AccountId, investment_id: Self::InvestmentId ) -> Result<Self::Amount, Self::Error>

Returns, if possible, the currently unprocessed investment amount (in pool currency) of who into the given investment class.

NOTE: If the investment was (partially) processed, the unprocessed amount is only updated upon collecting.

fn redemption( who: &AccountId, investment_id: Self::InvestmentId ) -> Result<Self::Amount, Self::Error>

Returns, if possible, the currently unprocessed redemption amount (in tranche tokens) of who into the given investment class.

NOTE: If the redemption was (partially) processed, the unprocessed amount is only updated upon collecting.

fn accepted_payment_currency( investment_id: Self::InvestmentId, currency: Self::CurrencyId ) -> bool

Checks whether a currency can be used for buying the given investment.

fn accepted_payout_currency( investment_id: Self::InvestmentId, currency: Self::CurrencyId ) -> bool

Checks whether a currency is accepted as a payout for the given investment.