Trait cfg_traits::investments::Investment

pub trait Investment<AccountId> {
    type Amount;
    type CurrencyId;
    type Error: Debug;
    type InvestmentId;

    // Required methods
    fn update_investment(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        amount: Self::Amount
    ) -> Result<(), Self::Error>;
    fn accepted_payment_currency(
        investment_id: Self::InvestmentId,
        currency: Self::CurrencyId
    ) -> bool;
    fn investment(
        who: &AccountId,
        investment_id: Self::InvestmentId
    ) -> Result<Self::Amount, Self::Error>;
    fn update_redemption(
        who: &AccountId,
        investment_id: Self::InvestmentId,
        amount: Self::Amount
    ) -> Result<(), Self::Error>;
    fn accepted_payout_currency(
        investment_id: Self::InvestmentId,
        currency: Self::CurrencyId
    ) -> bool;
    fn redemption(
        who: &AccountId,
        investment_id: Self::InvestmentId
    ) -> Result<Self::Amount, Self::Error>;
    fn investment_requires_collect(
        investor: &AccountId,
        investment_id: Self::InvestmentId
    ) -> bool;
    fn redemption_requires_collect(
        investor: &AccountId,
        investment_id: Self::InvestmentId
    ) -> bool;
}

A trait, when implemented allows to invest into investment classes

Required Associated Types

type Amount

type CurrencyId

type Error: Debug

type InvestmentId

Required Methods

fn update_investment( who: &AccountId, investment_id: Self::InvestmentId, amount: Self::Amount ) -> Result<(), Self::Error>

Updates the current investment amount of who into the investment class to amount. Meaning: if amount < previous investment, then investment will be reduced, and increases in the opposite case.

fn accepted_payment_currency( investment_id: Self::InvestmentId, currency: Self::CurrencyId ) -> bool

Checks whether a currency can be used for buying the given investment.

fn investment( who: &AccountId, investment_id: Self::InvestmentId ) -> Result<Self::Amount, Self::Error>

Returns, if possible, the currently unprocessed investment amount (in pool currency) of who into the given investment class.

NOTE: If the investment was (partially) processed, the unprocessed amount is only updated upon collecting.

fn update_redemption( who: &AccountId, investment_id: Self::InvestmentId, amount: Self::Amount ) -> Result<(), Self::Error>

Updates the current redemption amount (in tranche tokens) of who into the investment class to amount. Meaning: if amount < previous redemption, then the redemption will be reduced, and increased in the opposite case.

NOTE: Redemptions are bound by the processed investment amount.

fn accepted_payout_currency( investment_id: Self::InvestmentId, currency: Self::CurrencyId ) -> bool

Checks whether a currency is accepted as a payout for the given investment.

fn redemption( who: &AccountId, investment_id: Self::InvestmentId ) -> Result<Self::Amount, Self::Error>

Returns, if possible, the currently unprocessed redemption amount (in tranche tokens) of who into the given investment class.

NOTE: If the redemption was (partially) processed, the unprocessed amount is only updated upon collecting.

fn investment_requires_collect( investor: &AccountId, investment_id: Self::InvestmentId ) -> bool

Checks whether an investment requires to be collected before it can be updated.

NOTE: Defaults to false if the investment does not exist.

fn redemption_requires_collect( investor: &AccountId, investment_id: Self::InvestmentId ) -> bool

Checks whether a redemption requires to be collected before it can be further updated.

NOTE: Defaults to false if the redemption does not exist.