Macro cfg_traits::impl_mock_accountant

macro_rules! impl_mock_accountant {
    ($name:ident, $account_id:ty, $investment_id:ty, $currency_id:ty, $balance:ty) => { ... };
}

Exposes a struct $name that implements the trait Accountant. The struct expects one generic parameter that implements the fungibles traits Inspect, Mutate and Transfer. Furthermore, there exists a struct GenesisConfig that implements trait GenesisBuild that can be used like any other GenesisConfig to initialize state in the TestExternalities.

Also exports a struct InvestmentInfo to be used in the GenesisConfig

E.g.: MockAccountant<Tokens: frame_support::traits::tokens::fungibles::{Inspect, Mutate, Transfer}>

Example macro usage:

use cfg_traits::impl_mock_accountant;
use cfg_primitives::{PoolId, TrancheId, Balance};
use cfg_types::CurrencyId;
use frame_support::traits::fungibles::{Inspect, Mutate, Transfer};
use frame_support::traits::GenesisBuild;

/// The used account id for this mock
type AccountId = u64;

enum InvestmentId {
    Tranches(PoolId, TrancheId),
}

impl Into<CurrencyId> for InvestmentId {
	fn into(self) -> CurrencyId {
   		CurrencyId::Tranche(self.0, self.1)
    }
}

impl_mock_accountant!(
	MockAccountant,
    AccountId,
    InvestmentId,
    CurrencyId,
    Balance
);

// Using the `GenesisConfig`
use accountant_mock::InvestmentInfo;
let storage = GenesisBuild::build_storage(&accountant_mock::GenesisConfig {
		infos: vec![
			(
				InvestmentId::Tranche(0, [0;16]),
				accountant_mock::InvestmentInfo {
					owner: 1,
					id: InvestmentId::Tranches(0, [0;16]),
					payment_currency:CurrencyId::AUSD
				}
			)
		]
}).expect("Must not fail");