Module cfg_traits::ops::ensure

Arithmetic operations with safe error handling.

This module provide a more readable way to do safe arithmetics, turning this:

self.my_value = self.my_value.checked_sub(other_value).ok_or(ArithmeticError::Overflow)?;

into this:

self.my_value.ensure_sub_assign(other_value)?;

And choose the correct ArithmeticError it should return in case of fail.

The EnsureOps family functions follows the same behavior as CheckedOps but returning an ArithmeticError instead of None.

Traits

EnsureAdd

Performs addition that returns ArithmeticError instead of wrapping around on overflow.

EnsureAddAssign

Performs self addition that returns ArithmeticError instead of wrapping around on overflow.

EnsureDiv

Performs division that returns ArithmeticError instead of wrapping around on overflow.

EnsureDivAssign

Performs self division that returns ArithmeticError instead of wrapping around on overflow.

EnsureFixedPointNumber

Extends FixedPointNumber with the Ensure family functions.

EnsureFrom

Similar to TryFrom but returning an ArithmeticError error.

EnsureInto

Similar to TryInto but returning an ArithmeticError error.

EnsureMul

Performs multiplication that returns ArithmeticError instead of wrapping around on overflow.

EnsureMulAssign

Performs self multiplication that returns ArithmeticError instead of wrapping around on overflow.

EnsureSub

Performs subtraction that returns ArithmeticError instead of wrapping around on underflow.

EnsureSubAssign

Performs self subtraction that returns ArithmeticError instead of wrapping around on underflow.