Trait cfg_traits::InterestAccrual

pub trait InterestAccrual<InterestRate, Balance, Adjustment> {
    type NormalizedDebt: Member + Parameter + MaxEncodedLen + TypeInfo + Copy + Zero;

    fn current_debt(
        interest_rate_per_sec: InterestRate, 
        normalized_debt: Self::NormalizedDebt
    ) -> Result<Balance, DispatchError>;
    fn previous_debt(
        interest_rate_per_sec: InterestRate, 
        normalized_debt: Self::NormalizedDebt, 
        when: Moment
    ) -> Result<Balance, DispatchError>;
    fn adjust_normalized_debt(
        interest_rate_per_sec: InterestRate, 
        normalized_debt: Self::NormalizedDebt, 
        adjustment: Adjustment
    ) -> Result<Self::NormalizedDebt, DispatchError>;
    fn renormalize_debt(
        old_interest_rate: InterestRate, 
        new_interest_rate: InterestRate, 
        normalized_debt: Self::NormalizedDebt
    ) -> Result<Self::NormalizedDebt, DispatchError>;
    fn reference_yearly_rate(
        interest_rate_per_year: InterestRate
    ) -> Result<InterestRate, DispatchError>;
    fn reference_rate(interest_rate_per_sec: InterestRate) -> DispatchResult;
    fn unreference_rate(interest_rate_per_sec: InterestRate) -> DispatchResult;
    fn convert_additive_rate_to_per_sec(
        interset_rate_per_year: InterestRate
    ) -> Result<InterestRate, DispatchError>;
}

A trait that can be used to calculate interest accrual for debt

Required Associated Types

type NormalizedDebt: Member + Parameter + MaxEncodedLen + TypeInfo + Copy + Zero

Required Methods

fn current_debt(
interest_rate_per_sec: InterestRate,
normalized_debt: Self::NormalizedDebt
) -> Result<Balance, DispatchError>

Calculate the current debt using normalized debt * cumulative rate

fn previous_debt(
interest_rate_per_sec: InterestRate,
normalized_debt: Self::NormalizedDebt,
when: Moment
) -> Result<Balance, DispatchError>

Calculate a previous debt using normalized debt * previous cumulative rate

If when is further in the past than the last time the normalized debt was adjusted, this will return nonsense (effectively “rewinding the clock” to before the value was valid)

fn adjust_normalized_debt(
interest_rate_per_sec: InterestRate,
normalized_debt: Self::NormalizedDebt,
adjustment: Adjustment
) -> Result<Self::NormalizedDebt, DispatchError>

Increase or decrease the normalized debt

fn renormalize_debt(
old_interest_rate: InterestRate,
new_interest_rate: InterestRate,
normalized_debt: Self::NormalizedDebt
) -> Result<Self::NormalizedDebt, DispatchError>

Re-normalize a debt for a new interest rate

fn reference_yearly_rate(
interest_rate_per_year: InterestRate
) -> Result<InterestRate, DispatchError>

Indicate that a yearly rate is in use

Validates that the rate is allowed, and converts it to a per-second rate for future operations

fn reference_rate(interest_rate_per_sec: InterestRate) -> DispatchResult

Indicate that a rate is in use

fn unreference_rate(interest_rate_per_sec: InterestRate) -> DispatchResult

Indicate that a rate is no longer in use

fn convert_additive_rate_to_per_sec(
interset_rate_per_year: InterestRate
) -> Result<InterestRate, DispatchError>

Verifies a yearly additive rate and converts it to a per-second additive rate