Trait cfg_traits::Investment

pub trait Investment<AccountId> {
    type Error;
    type InvestmentId;
    type Amount;

    fn update_investment(
        who: &AccountId, 
        investment_id: Self::InvestmentId, 
        amount: Self::Amount
    ) -> Result<(), Self::Error>;
    fn investment(
        who: &AccountId, 
        investment_id: Self::InvestmentId
    ) -> Result<Self::Amount, Self::Error>;
    fn update_redemption(
        who: &AccountId, 
        investment_id: Self::InvestmentId, 
        amount: Self::Amount
    ) -> Result<(), Self::Error>;
    fn redemption(
        who: &AccountId, 
        investment_id: Self::InvestmentId
    ) -> Result<Self::Amount, Self::Error>;
}

A trait, when implemented allows to invest into investment classes

Required Associated Types

type Error

type InvestmentId

type Amount

Required Methods

fn update_investment(
who: &AccountId,
investment_id: Self::InvestmentId,
amount: Self::Amount
) -> Result<(), Self::Error>

Updates the current investment amount of who into the investment class to amount. Meaning: if amount < previous investment, then investment will be reduced, and increases in the opposite case.

fn investment(
who: &AccountId,
investment_id: Self::InvestmentId
) -> Result<Self::Amount, Self::Error>

Returns, if possible, the current investment amount of who into the given investment class

fn update_redemption(
who: &AccountId,
investment_id: Self::InvestmentId,
amount: Self::Amount
) -> Result<(), Self::Error>

Updates the current redemption amount of who into the investment class to amount. Meaning: if amount < previous redemption, then redemption will be reduced, and increases in the opposite case.

fn redemption(
who: &AccountId,
investment_id: Self::InvestmentId
) -> Result<Self::Amount, Self::Error>

Returns, if possible, the current redemption amount of who into the given investment class