Trait cfg_traits::InvestmentProperties

pub trait InvestmentProperties<AccountId> {
    type Currency;
    type Id;

    fn owner(&self) -> AccountId;
    fn id(&self) -> Self::Id;
    fn payment_currency(&self) -> Self::Currency;

    fn payment_account(&self) -> AccountId { ... }
}

A trait that allows to retrieve information about an investment class.

Required Associated Types

type Currency

The overarching Currency that payments for this class are made in

type Id

Who the investment class can be identified

Required Methods

fn owner(&self) -> AccountId

Returns the owner of the investment class

fn id(&self) -> Self::Id

Returns the id of the investment class

fn payment_currency(&self) -> Self::Currency

Returns the currency in which the investment class can be bought.

Provided Methods

fn payment_account(&self) -> AccountId

Returns the account a payment for the investment class must be made to.

Defaults to owner.