Trait cfg_traits::OrderManager

pub trait OrderManager {
    type Error;
    type InvestmentId;
    type Orders;
    type Fulfillment;

    fn invest_orders(
        asset_id: Self::InvestmentId
    ) -> Result<Self::Orders, Self::Error>;
    fn redeem_orders(
        asset_id: Self::InvestmentId
    ) -> Result<Self::Orders, Self::Error>;
    fn invest_fulfillment(
        asset_id: Self::InvestmentId, 
        fulfillment: Self::Fulfillment
    ) -> Result<(), Self::Error>;
    fn redeem_fulfillment(
        asset_id: Self::InvestmentId, 
        fulfillment: Self::Fulfillment
    ) -> Result<(), Self::Error>;
}

A trait, when implemented must take care of collecting orders (invest & redeem) for a given investment class. When being asked it must return the current orders and when being singled about a fulfillment, it must act accordingly.

Required Associated Types

type Error

type InvestmentId

type Orders

type Fulfillment

Required Methods

fn invest_orders(
asset_id: Self::InvestmentId
) -> Result<Self::Orders, Self::Error>

When called the manager return the current invest orders for the given investment class.

fn redeem_orders(
asset_id: Self::InvestmentId
) -> Result<Self::Orders, Self::Error>

When called the manager return the current redeem orders for the given investment class.

fn invest_fulfillment(
asset_id: Self::InvestmentId,
fulfillment: Self::Fulfillment
) -> Result<(), Self::Error>

Signals the manager that the previously fetch invest orders for a given investment class will be fulfilled by fulfillment.

fn redeem_fulfillment(
asset_id: Self::InvestmentId,
fulfillment: Self::Fulfillment
) -> Result<(), Self::Error>

Signals the manager that the previously fetch redeem orders for a given investment class will be fulfilled by fulfillment.