Trait cfg_traits::PoolInspect

pub trait PoolInspect<AccountId, CurrencyId> {
    type PoolId: Parameter + Member + Debug + Copy + Default + TypeInfo;
    type TrancheId: Parameter + Member + Debug + Copy + Default + TypeInfo;
    type Rate;
    type Moment;

    fn pool_exists(pool_id: Self::PoolId) -> bool;
    fn tranche_exists(
        pool_id: Self::PoolId, 
        tranche_id: Self::TrancheId
    ) -> bool;
    fn get_tranche_token_price(
        pool_id: Self::PoolId, 
        tranche_id: Self::TrancheId
    ) -> Option<PriceValue<CurrencyId, Self::Rate, Self::Moment>>;
}

A trait that support pool inspection operations such as pool existence checks and pool admin of permission set.

Required Associated Types

type PoolId: Parameter + Member + Debug + Copy + Default + TypeInfo

type TrancheId: Parameter + Member + Debug + Copy + Default + TypeInfo

type Rate

type Moment

Required Methods

fn pool_exists(pool_id: Self::PoolId) -> bool

check if the pool exists

fn tranche_exists(pool_id: Self::PoolId, tranche_id: Self::TrancheId) -> bool

fn get_tranche_token_price(
pool_id: Self::PoolId,
tranche_id: Self::TrancheId
) -> Option<PriceValue<CurrencyId, Self::Rate, Self::Moment>>