Trait cfg_traits::TokenSwaps

pub trait TokenSwaps<Account> {
    type CurrencyId;
    type Balance;
    type SellRatio;
    type OrderId;
    type OrderDetails;

    // Required methods
    fn place_order(
        account: Account,
        currency_in: Self::CurrencyId,
        currency_out: Self::CurrencyId,
        buy_amount: Self::Balance,
        sell_rate_limit: Self::SellRatio
    ) -> Result<Self::OrderId, DispatchError>;
    fn update_order(
        account: Account,
        order_id: Self::OrderId,
        buy_amount: Self::Balance,
        sell_rate_limit: Self::SellRatio
    ) -> DispatchResult;
    fn valid_pair(
        currency_in: Self::CurrencyId,
        currency_out: Self::CurrencyId
    ) -> bool;
    fn cancel_order(order: Self::OrderId) -> DispatchResult;
    fn is_active(order: Self::OrderId) -> bool;
    fn get_order_details(order: Self::OrderId) -> Option<Self::OrderDetails>;
}

Required Associated Types

type CurrencyId

type Balance

type SellRatio

type OrderId

type OrderDetails

Required Methods

fn place_order( account: Account, currency_in: Self::CurrencyId, currency_out: Self::CurrencyId, buy_amount: Self::Balance, sell_rate_limit: Self::SellRatio ) -> Result<Self::OrderId, DispatchError>

Swap tokens buying a buy_amount of currency_in using the currency_out tokens. The implementer of this method should know the current market rate between those two currencies. sell_rate_limit defines the highest price acceptable for currency_in currency when buying with currency_out. This protects order placer if market changes unfavourably for swap order. For example, with a sell_rate_limit of 3/2, one asset_in should never cost more than 1.5 units of asset_out. Returns Result with OrderId upon successful order creation.

NOTE: The minimum fulfillment amount is implicitly set by the implementor.

Example usage with pallet_order_book impl:

OrderBook::place_order(
    {AccountId},
    CurrencyId::ForeignAsset(0),
    CurrencyId::ForeignAsset(1),
    100 * FOREIGN_ASSET_0_DECIMALS,
    Quantity::checked_from_rational(3u32, 2u32).unwrap(),
    100 * FOREIGN_ASSET_0_DECIMALS
)

Would return Ok({OrderId} and create the following order in storage:

Order {
    order_id: {OrderId},
    placing_account: {AccountId},
    asset_in_id: CurrencyId::ForeignAsset(0),
    asset_out_id: CurrencyId::ForeignAsset(1),
    buy_amount: 100 * FOREIGN_ASSET_0_DECIMALS,
    initial_buy_amount: 100 * FOREIGN_ASSET_0_DECIMALS,
    sell_rate_limit: Quantity::checked_from_rational(3u32, 2u32).unwrap(),
    max_sell_amount: 150 * FOREIGN_ASSET_1_DECIMALS,
    min_fulfillment_amount: 10 * CFG * FOREIGN_ASSET_0_DECIMALS,
}

fn update_order( account: Account, order_id: Self::OrderId, buy_amount: Self::Balance, sell_rate_limit: Self::SellRatio ) -> DispatchResult

Update an existing active order. As with creating an order, the sell_rate_limit defines the highest price acceptable for currency_in currency when buying with currency_out. Returns a Dispatch result.

NOTE: The minimum fulfillment amount is implicitly set by the implementor.

This Can fail for various reasons.

Example usage with pallet_order_book impl:

OrderBook::update_order(
    {AccountId},
    {OrderId},
    15 * FOREIGN_ASSET_0_DECIMALS,
    Quantity::checked_from_integer(2u32).unwrap(),
    6 * FOREIGN_ASSET_0_DECIMALS
)

Would return Ok(()) and update the following order in storage:

Order {
    order_id: {OrderId},
    placing_account: {AccountId},
    asset_in_id: CurrencyId::ForeignAsset(0),
    asset_out_id: CurrencyId::ForeignAsset(1),
    buy_amount: 15 * FOREIGN_ASSET_0_DECIMALS,
    initial_buy_amount: 100 * FOREIGN_ASSET_0_DECIMALS,
    sell_rate_limit: Quantity::checked_from_integer(2u32).unwrap(),
    max_sell_amount: 30 * FOREIGN_ASSET_1_DECIMALS
    min_fulfillment_amount: 10 * CFG * FOREIGN_ASSET_0_DECIMALS,
}

fn valid_pair( currency_in: Self::CurrencyId, currency_out: Self::CurrencyId ) -> bool

A sanity check that can be used for validating that a trading pair is supported. Will also be checked when placing an order but might be cheaper.

fn cancel_order(order: Self::OrderId) -> DispatchResult

Cancel an already active order.

fn is_active(order: Self::OrderId) -> bool

Check if the order is still active.

fn get_order_details(order: Self::OrderId) -> Option<Self::OrderDetails>

Retrieve the details of the order if it exists.