Crate cfg_traits

A common trait lib for centrifuge

This crate provides some common traits used by centrifuge.

Modules

changes

Traits related to checked changes.

data

Traits related to data registry and collection.

ethereum

Traits related to Ethereum/EVM.

fees

interest

Traits related to interest rates.

investments

Traits related to investments.

liquidity_pools

Traits related to liquidity pools.

rewards

Traits related to rewards.

Structs

Always

CurrencyPair

A trait that can be used to retrieve the current price for a currency

Never

PriceValue

Enums

UpdateState

Variants for valid Pool updates to send out as events

Traits

CurrencyInspect

Trait to retrieve information about currencies.

CurrencyPrice

IdentityCurrencyConversion

Trait to synchronously provide a currency conversion estimation for foreign currencies into/from pool currencies.

Permissions

PoolInspect

A trait that support pool inspection operations such as pool existence checks and pool admin of permission set.

PoolMetadata

A trait that supports retrieval and mutation of pool and tranche token metadata.

PoolMutate

A trait that supports modifications of pools

PoolNAV

A trait that can be used to fetch the nav and update nav for a given pool

PoolReserve

A trait that support pool reserve operations such as withdraw and deposit

PoolUpdateGuard

PoolWriteOffPolicyMutate

A trait that supports modifications of pool write-off policies

PreConditions

Properties

Reward

A trait used for loosely coupling the claim pallet with a reward mechanism.

StatusNotificationHook

Trait to transmit a change of status for anything uniquely identifiable.

TokenSwaps

TrancheTokenPrice

Get the latest price for a given tranche token

TransferAllowance

Trait to determine whether a sending account and currency have a restriction, and if so is there an allowance for the receiver location.

TryConvert

A trait for trying to convert between two types.