Trait cfg_traits::Reward

pub trait Reward {
    type ParachainAccountId: Debug + MaybeSerialize + MaybeSerializeDeserialize + Member + Ord + Parameter + TypeInfo;
    type ContributionAmount: AtLeast32BitUnsigned + Codec + Copy + Debug + Default + MaybeSerializeDeserialize + Member + Parameter + Zero + TypeInfo;
    type BlockNumber: AtLeast32BitUnsigned + Bounded + Copy + Debug + Default + FromStr + Hash + MaybeDisplay + MaybeSerializeDeserialize + Member + Parameter + TypeInfo;

    // Required method
    fn reward(
        who: Self::ParachainAccountId,
        contribution: Self::ContributionAmount
    ) -> DispatchResultWithPostInfo;
}

A trait used for loosely coupling the claim pallet with a reward mechanism.

Overview

The crowdloan reward mechanism is separated from the crowdloan claiming process, the latter being generic, acting as a kind of proxy to the rewarding mechanism, that is specific to to each crowdloan campaign. The aim of this pallet is to ensure that a claim for a reward payout is well-formed, checking for replay attacks, spams or invalid claim (e.g. unknown contributor, exceeding reward amount, …). See the [crowdloan-reward] pallet, that implements a reward mechanism with vesting, for instance.

Required Associated Types

type ParachainAccountId: Debug + MaybeSerialize + MaybeSerializeDeserialize + Member + Ord + Parameter + TypeInfo

The account from the parachain, that the claimer provided in her/his transaction.

type ContributionAmount: AtLeast32BitUnsigned + Codec + Copy + Debug + Default + MaybeSerializeDeserialize + Member + Parameter + Zero + TypeInfo

The contribution amount in relay chain tokens.

type BlockNumber: AtLeast32BitUnsigned + Bounded + Copy + Debug + Default + FromStr + Hash + MaybeDisplay + MaybeSerializeDeserialize + Member + Parameter + TypeInfo

Block number type used by the runtime

Required Methods

fn reward( who: Self::ParachainAccountId, contribution: Self::ContributionAmount ) -> DispatchResultWithPostInfo

Rewarding function that is invoked from the claim pallet.

If this function returns successfully, any subsequent claim of the same claimer will be rejected by the claim module.